requirements.txt
accelerate==0.21.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
diffusers==0.20.0
filelock==3.9.0
fonttools==4.42.0
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
invisible-watermark==0.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
mpmath==1.2.1
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.1
opencv-python==4.8.0.74
packaging==23.1
peft==0.4.0
Pillow==9.3.0
psutil==5.9.5
pyparsing==3.0.9
pyre-extensions==0.0.29
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.6.3
requests==2.28.1
safetensors==0.3.1
six==1.16.0
style==1.1.0
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.65.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.4.0
update==0.0.1
urllib3==1.26.13
xformers==0.0.20
zipp==3.16.2
